MoreExercise - 01. ExtractPersonInformation
1.	Extract Person Information
Write a program that reads N lines of strings and extracts the name and age of a given person. The name of the person will be between '@' and '|'. The person’s age will be between '#' and '*'. Example: "Hello my name is @Peter| and I am #20* years old." For each found name and age print a line in the following format "{name} is {age} years old."